Add some idiomatic functionality
 - Move foreign function interface part to separate module to indicate
   its origin.

 - Add sketch of DRM functionality.
