+++ src/ffi/mod.rs
@@ -1,0 +1,14 @@
+#[allow(dead_code)]
+pub mod util;
+
+#[allow(dead_code, non_camel_case_types)]
+pub mod drm_mode;
+
+#[allow(dead_code, non_camel_case_types)]
+pub mod drm;
+
+#[allow(dead_code, non_camel_case_types)]
+pub mod xf86drm;
+
+#[allow(dead_code, non_camel_case_types, non_snake_case)]
+pub mod xf86drm_mode;
