+++ src/ffi/util.rs
