+++ src/lib.rs
@@ -3,16 +3,13 @@ extern crate libc;
 extern crate byteorder;
 
 #[allow(dead_code)]
-pub mod util;
+pub mod ffi;
 
-#[allow(dead_code, non_camel_case_types)]
-pub mod drm_mode;
-
-#[allow(dead_code, non_camel_case_types)]
+#[allow(dead_code)]
 pub mod drm;
 
-#[allow(dead_code, non_camel_case_types)]
-pub mod xf86drm;
+#[allow(dead_code)]
+pub mod drm_mode;
 
-#[allow(dead_code, non_camel_case_types, non_snake_case)]
-pub mod xf86drm_mode;
+#[allow(dead_code)]
+mod resources;
